return parent rows in searches, with tests
